refactor: only have `use client` for custom element
